Give warnings for duplicate declarations and list conflicting modules
Provide warnings when a duplicate declaration is detected to include
where the two conflicting declarations occurred.

When the duplicate declarations are in different modules, add the two
modules to a table of conflicting modules.

Create the new lua module refpolicy_modules_conflicting with the
function list_conflicting_modules() to warn about the conflicting
modules. Have spt_ling.lua call this function after other processing
is complete.

Signed-off-by: James Carter <jwcart2@tycho.nsa.gov>