Determine and display list of module dependencies
When checking statements and an identifier is found that is not
declared in the module, add the module where the declaration is
made to a table of dependencies for the module being checked.

Also, when a macro call is found that is outside of an optional
block then add the module that defines the macro to the table
of dependencies for the module being checked.

Create the new lua module refpolicy_modules_dependencies with the
function list_dependent_modules() to warn about various types of
dependencies. The most serious dependency is when a base module
depdends on a non-base module. Other dependencies, such as a contrib
module depdending on another contrib module, may or may not indicate
a problem.

Signed-off-by: James Carter <jwcart2@tycho.nsa.gov>