Report deprecated and unimplemented macros
Report calls to deprecated macros so that they can be updated to
call the new macro. Report calls to unimplemented macros (which
perhaps is serving as a placeholder until the policy can be
completed).

Add a flags field to the macro definition node and update the flags
when a refpolicywarn() macro is found. Create the new lua module
refpolicy_macros_avoid containing the function
check_for_macros_to_avoid() which reports the deprecated and
unimplemented macros.

Signed-off-by: James Carter <jwcart2@tycho.nsa.gov>