Change how gen_user() is parsed
Before mls_level, mls_range, and mcs_cats was saved, but it would
be hard to translate what was saved into a new policy language.

Now create enable_mls and enable_mcs ifdef blocks and create user
rule for non-mls, mls, and mcs (if mcs_cats is specified) in those
blocks. This makes it easy to translate into a new policy language.

Signed-off-by: James Carter <jwcart2@tycho.nsa.gov>