Added c function to create a new directory
Created a new lua function in the selpoltools c library that creates a
new directory given the pathname of the new directory.

All of the directories prior to the final one must have already been
created.

The function will return true if successful and false along with an
error message if it is not successful.

Signed-off-by: James Carter <jwcart2@tycho.nsa.gov>