lua: Add support for a file type in genfscon rules
Although rarely used, genfscon rules support the specification of a
file type just like the rules in a file context file. The file type
is used to make the genfscon rule apply only for a specific security
class.

Support parsing genfscon rules with an optional file type and write
out the file type, if specified, when writing genfscon rules.

Signed-off-by: James Carter <jwcart2@gmail.com>