lua: Handle m4 ifelse when used like a case statement
An ifelse conditional can behave like a case statement.
ifelse(`S1',`S2',`
  EQUAL-1
  . . .
',`S3',`S4',`
  EQUAL-2
  . . .
',`S5',`S6',`
  EQUAL-3
  . . .
',`
  NOT-EQUAL-4
  . . .
')

EQUAL-1 is expanded if S1 and S2 expand to the same thing. Otherwise,
EQUAL-2 is expanded if S3 and S4 expand to the same thing. Otherwise,
EQUAL-3 is expanded if S5 and S6 expand to the same thing. Otherwise,
NOT-EQUAL-4 is expanded.

This behavior occurs when the ifelse has at least six arguments.
There is no set maximum number of arguments.

Signed-off-by: James Carter <jwcart2@gmail.com>