lua: Add functions to add to and write a warnings buffer
The function warnings_buffer_add() is used to add a warning
message to a warnings buffer. The function warnings_buffer_write()
sorts and writes out the warnings in the given buffer. The functions
warnings_buffer_write1(), warnings_buffer_write2(), and
warnings_buffer_write3() write out the given buffer if verbose is
greater than 0, 1, or 2 respectively.

Signed-off-by: James Carter <jwcart2@gmail.com>