docs and lua: Change warning for call outside of an optional block
There are a lot of warning for macro calls outside of optional
blocks. Mostly these are going to be harmless because they are to
macros declared in modules that are always loaded. To reduce the
number of warnings for verbose < 3, split out the warnings for
calls in base modules, which should always be displayed, from the
warnings for calls in non-base modules, which should only be
displayed when verbose == 3.

Signed-off-by: James Carter <jwcart2@gmail.com>