lua/refpolicy_parse.lua: Fix minor errors in creating parse tree
Make sure that all variables being saved in the node data have a
value and are not nil. This means that optional data in a rule
should be set to false.

Type change and type member rules were using get_identifier_or_set()
for the new type, but a set is not allowed for the new type, so
just use get_identifier().

When setting the data for a macro call, extra data was passed to
MACRO.set_call_data(). Only pass the data that is needed.

Signed-off-by: James Carter <jwcart2@gmail.com>